third_party/WebKit/LayoutTests/push_messaging/push-messaging-has-permission-expected.txt

Tests for the Push API hasPermission method.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS status is "default"
PASS status is "denied"
PASS status is "granted"
PASS successfullyParsed is true

TEST COMPLETE
